perf: extends를 사용한 성능 개선
