--- BE_homework_/src/SpaceshipType.java
@@ -1,12 +1,0 @@
-public enum SpaceshipType {
-    MANNED(true), UNMANNED(false);
-    private final boolean isManned;
-
-    SpaceshipType(boolean isManned) {
-        this.isManned=isManned;
-    }
-
-    public boolean isManned(){
-        return isManned;
-    }
-}
